requirements.txt
Django==5.0.7
django-jazzmin==3.0.0
djangorestframework==3.15.2
drf-yasg==1.21.7
pillow==10.4.0